+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # ui
 
+## 1.9.2-next.14
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.9.2-next.14
+- @copilotkit/shared@1.9.2-next.14
+
+## 1.9.2-next.13
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.9.2-next.13
+- @copilotkit/shared@1.9.2-next.13
+
 ## 1.9.2-next.12
 
 ### Patch Changes