+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkit/runtime-client-gql
 
+## 1.9.2-next.14
+
+### Patch Changes
+
+- @copilotkit/shared@1.9.2-next.14
+
+## 1.9.2-next.13
+
+### Patch Changes
+
+- @copilotkit/shared@1.9.2-next.13
+
 ## 1.9.2-next.12
 
 ### Patch Changes