+++ CopilotKit/packages/sdk-js/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @copilotkit/sdk-js
 
+## 1.9.2-next.14
+
+### Patch Changes
+
+- @copilotkit/shared@1.9.2-next.14
+
+## 1.9.2-next.13
+
+### Patch Changes
+
+- @copilotkit/shared@1.9.2-next.13
+
 ## 1.9.2-next.12
 
 ### Patch Changes